(clickhouse) feat: update table for cluster (#2361)
