+++ jobs/etl_jobs/internal/export_clickhouse/schema/analytics/daily_aggregated_offer_consultation.sql
@@ -1,0 +1,12 @@
+CREATE VIEW analytics.daily_aggregated_offer_consultation ON cluster default
+AS
+SELECT
+    toDate(partition_date) AS event_date,
+    origin,
+    offer_id,
+    sum(cnt_events) AS sumCountViews
+FROM intermediate.offer_consultation
+GROUP BY
+    event_date,
+    origin,
+    offer_id
