+++ jobs/etl_jobs/internal/export_clickhouse/schema/intermediate/user_consultation.sql
@@ -1,4 +1,4 @@
-CREATE TABLE IF NOT EXISTS {{ dataset }}.user_consultation(
+CREATE TABLE IF NOT EXISTS intermediate.user_consultation ON cluster default(
     partition_date String,
     event_timestamp Datetime64,
     offer_id String, 
@@ -8,5 +8,6 @@ CREATE TABLE IF NOT EXISTS {{ dataset }}.user_consultation(
   ENGINE = MergeTree
   PARTITION BY partition_date
   ORDER BY tuple(user_id, offer_id)
+  SETTINGS storage_policy='gcs_main'
   COMMENT 'Users consultations on native app, partitioned by date'
