+++ orchestration/dags/dependencies/export_clickhouse/sql/export_user_consultation.sql
@@ -6,7 +6,7 @@ SELECT
     origin
 FROM `{{ bigquery_analytics_dataset }}.firebase_events` 
 WHERE 
-event_date BETWEEN date_sub(DATE("{{ ds }}"), INTERVAL 4 DAY) and DATE("{{ ds }}")
+event_date BETWEEN date_sub(DATE("{{ ds }}"), INTERVAL {{ params.days }} DAY) and DATE("{{ ds }}")
 and event_name = "ConsultOffer"
 AND offer_id is not null
 AND user_id is not null
